requirements.txt
Flask
requests
gunicorn
newsapi-python